SAE_formularios_ms/requirements.txt
fastapi
uvicorn
pymongo==4.3.3
python-dotenv==1.0.0